requirements.txt
pytest-playwright
pytest-dotenv
pandas==1.5.0
supabase==0.7.1